Make Creature a subclass of Entity
