+++ pythonlike/src/game_map.py
@@ -1,16 +1,16 @@
 import numpy as np
-from entity import Entity
-
+from creature import Creature
+import entity
 
 class GameMap:
-    def __init__(self, w: int, h: int, tiles: np.ndarray, entities: list[Entity]):
+    def __init__(self, w: int, h: int, tiles: np.ndarray, entities: list[entity.Entity]):
         self.w = w
         self.h = h
         self.tiles = tiles
-        self.entities = entities
+        self.creatures = entities
         self.player = entities[0]  # Player should always be first when initializing creatures
-        # self.objects = []
-        # self.items = []
+        self.objects = []
+        self.items = []
         self.explored = np.zeros_like(self.tiles)
         self.visible = np.zeros_like(self.tiles)
 
@@ -20,16 +20,16 @@ class GameMap:
     def is_occupied(self, x: int, y:int) -> bool:
         return True if self.get_creature(x, y) else False
 
-    def get_creature(self, x: int, y: int) -> Entity | None:
+    def get_creature(self, x: int, y: int) -> Creature | None:
         """Return creature at (x,y) or None if there is no creature."""
-        for e in self.entities:
+        for e in self.creatures:
             if e.x == x and e.y == y:
                 return e
         return None
     
     def get_creature_index(self, x: int, y: int) -> int | None:
         """Return index of entity at (x, y) in entities."""
-        for i, e in enumerate(self.entities):
+        for i, e in enumerate(self.creatures):
             if e.x == x and e.y == y:
                 return i
         return None
