+++ pythonlike/src/util.py
@@ -1,0 +1,11 @@
+from typing import TypeAlias
+import numpy as np
+
+
+GraphicType: TypeAlias = tuple[int, tuple[int, int, int, int], tuple[int, int, int, int]]
+
+graphic_dtype = np.dtype([
+    ("ch", np.intc),
+    ("fg", "(4,)u1"),
+    ("bg", "(4,)u1"),
+    ])
